Adding more stuff to insertion page
