+++ frontend/src/pages/NewExample/NewExample.css
